LVGL: Add support of LVGL9 into LVGL port
