+++ components/esp_lvgl_port/CMakeLists.txt
@@ -1,29 +1,61 @@
-file(GLOB_RECURSE IMAGE_SOURCES images/*.c)
 
-idf_component_register(SRCS "esp_lvgl_port.c" ${IMAGE_SOURCES} INCLUDE_DIRS "include" REQUIRES "esp_lcd" PRIV_REQUIRES "esp_timer")
+#Get LVGL version
+idf_component_get_property(lvgl_ver lvgl__lvgl COMPONENT_VERSION)
+message(STATUS "Component version: ${lvgl_ver}")
+
+#Select folder by LVGL version
+if(lvgl_ver LESS "9.0.0")
+    message(STATUS "Compiling LVGL8")
+    set(PORT_FOLDER "lvgl8")
+else()
+    message(STATUS "Compiling LVGL9")
+    set(PORT_FOLDER "lvgl9")
+endif()
+
+set(PORT_PATH "src/${PORT_FOLDER}")
+
+idf_component_register(SRCS "${PORT_PATH}/esp_lvgl_port.c" "${PORT_PATH}/esp_lvgl_port_disp.c" INCLUDE_DIRS "include" REQUIRES "esp_lcd" PRIV_REQUIRES "esp_timer")
+
+set(ADD_SRCS "")
+set(ADD_LIBS "")
 
 idf_build_get_property(build_components BUILD_COMPONENTS)
 if("espressif__button" IN_LIST build_components)
-    target_link_libraries(${COMPONENT_LIB} PRIVATE idf::espressif__button)
+    list(APPEND ADD_SRCS "${PORT_PATH}/esp_lvgl_port_button.c")
+    list(APPEND ADD_LIBS idf::espressif__button)
 endif()
 if("button" IN_LIST build_components)
-    target_link_libraries(${COMPONENT_LIB} PRIVATE idf::button)
+    list(APPEND ADD_SRCS "${PORT_PATH}/esp_lvgl_port_button.c")
+    list(APPEND ADD_LIBS idf::button)
 endif()
 if("espressif__esp_lcd_touch" IN_LIST build_components)
-    target_link_libraries(${COMPONENT_LIB} PRIVATE idf::espressif__esp_lcd_touch)
+    list(APPEND ADD_SRCS "${PORT_PATH}/esp_lvgl_port_touch.c")
+    list(APPEND ADD_LIBS idf::espressif__esp_lcd_touch)
 endif()
 if("esp_lcd_touch" IN_LIST build_components)
-    target_link_libraries(${COMPONENT_LIB} PRIVATE idf::esp_lcd_touch)
+    list(APPEND ADD_SRCS "${PORT_PATH}/esp_lvgl_port_touch.c")
+    list(APPEND ADD_LIBS idf::esp_lcd_touch)
 endif()
 if("espressif__knob" IN_LIST build_components)
-    target_link_libraries(${COMPONENT_LIB} PRIVATE idf::espressif__knob)
+    list(APPEND ADD_SRCS "${PORT_PATH}/esp_lvgl_port_knob.c")
+    list(APPEND ADD_LIBS idf::espressif__knob)
 endif()
 if("knob" IN_LIST build_components)
-    target_link_libraries(${COMPONENT_LIB} PRIVATE idf::knob)
+    list(APPEND ADD_SRCS "${PORT_PATH}/esp_lvgl_port_knob.c")
+    list(APPEND ADD_LIBS idf::knob)
 endif()
 if("espressif__usb_host_hid" IN_LIST build_components)
-    target_link_libraries(${COMPONENT_LIB} PRIVATE idf::espressif__usb_host_hid)
+    list(APPEND ADD_SRCS "${PORT_PATH}/esp_lvgl_port_usbhid.c" "images/${PORT_FOLDER}/img_cursor.c")
+    list(APPEND ADD_LIBS idf::espressif__usb_host_hid)
 endif()
 if("usb_host_hid" IN_LIST build_components)
-    target_link_libraries(${COMPONENT_LIB} PRIVATE idf::usb_host_hid)
+    list(APPEND ADD_SRCS "${PORT_PATH}/esp_lvgl_port_usbhid.c" "images/${PORT_FOLDER}/img_cursor.c")
+    list(APPEND ADD_LIBS idf::usb_host_hid)
+endif()
+
+if(ADD_SRCS)
+    target_sources(${COMPONENT_LIB} PRIVATE ${ADD_SRCS})
+endif()
+if(ADD_LIBS)
+    target_link_libraries(${COMPONENT_LIB} PRIVATE ${ADD_LIBS})
 endif()
