+++ components/esp_lvgl_port/README.md
@@ -13,6 +13,16 @@ This component helps with using LVGL with Espressif's LCD and touch drivers. It 
 * Add/remove navigation buttons input (using [`button`](https://github.com/espressif/esp-iot-solution/tree/master/components/button))
 * Add/remove encoder input (using [`knob`](https://github.com/espressif/esp-iot-solution/tree/master/components/knob))
 
+## LVGL Version
+
+This component supports **LVGL8** and **LVGL9**. By default, it selects the latest LVGL version. If you want to use a specific version (e.g. latest LVGL8), you can easily put into `idf_component.yml` in your project like this:
+
+```
+  lvgl/lvgl:
+    version: "^8"
+    public: true
+```
+
 ## Usage
 
 ### Initialization
@@ -53,6 +63,7 @@ Add an LCD screen to the LVGL. It can be called multiple times for adding multip
         },
         .flags = {
             .buff_dma = true,
+            .swap_bytes = false,
         }
     };
     disp_handle = lvgl_port_add_disp(&disp_cfg);
@@ -219,6 +230,7 @@ Every LVGL calls must be protected with these lock/unlock commands:
 ```
 
 ### Rotating screen
+
 LVGL port supports rotation of the display. You can select whether you'd like software rotation or hardware rotation.
 Software rotation requires no additional logic in your `flush_cb` callback.
 
@@ -233,6 +245,7 @@ Rotation mode can be selected in the `lvgl_port_display_cfg_t` structure.
     }
 ```
 Display rotation can be changed at runtime.
+
 ``` c
     lv_disp_set_rotation(disp_handle, LV_DISP_ROT_90);
 ```
