+++ components/esp_lvgl_port/examples/touchscreen/main/main.c
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2022-2023 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2022-2024 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -57,7 +57,7 @@ static esp_lcd_panel_handle_t lcd_panel = NULL;
 static esp_lcd_touch_handle_t touch_handle = NULL;
 
 /* LVGL display and touch */
-static lv_disp_t *lvgl_disp = NULL;
+static lv_display_t *lvgl_disp = NULL;
 static lv_indev_t *lvgl_touch_indev = NULL;
 
 static esp_err_t app_lcd_init(void)
@@ -190,6 +190,7 @@ static esp_err_t app_lvgl_init(void)
         },
         .flags = {
             .buff_dma = true,
+            .swap_bytes = true,
         }
     };
     lvgl_disp = lvgl_port_add_disp(&disp_cfg);
@@ -206,10 +207,10 @@ static esp_err_t app_lvgl_init(void)
 
 static void _app_button_cb(lv_event_t *e)
 {
-    lv_disp_rot_t rotation = lv_disp_get_rotation(lvgl_disp);
+    lv_disp_rotation_t rotation = lv_disp_get_rotation(lvgl_disp);
     rotation++;
-    if (rotation > LV_DISP_ROT_270) {
-        rotation = LV_DISP_ROT_NONE;
+    if (rotation > LV_DISPLAY_ROTATION_270) {
+        rotation = LV_DISPLAY_ROTATION_0;
     }
 
     /* LCD HW rotation */
@@ -227,10 +228,14 @@ static void app_main_display(void)
 
     /* Label */
     lv_obj_t *label = lv_label_create(scr);
-    lv_label_set_recolor(label, true);
     lv_obj_set_width(label, EXAMPLE_LCD_H_RES);
     lv_obj_set_style_text_align(label, LV_TEXT_ALIGN_CENTER, 0);
+#if LVGL_VERSION_MAJOR == 8
+    lv_label_set_recolor(label, true);
     lv_label_set_text(label, "#FF0000 "LV_SYMBOL_BELL" Hello world Espressif and LVGL "LV_SYMBOL_BELL"#\n#FF9400 "LV_SYMBOL_WARNING" For simplier initialization, use BSP "LV_SYMBOL_WARNING" #");
+#else
+    lv_label_set_text(label, LV_SYMBOL_BELL" Hello world Espressif and LVGL "LV_SYMBOL_BELL"\n "LV_SYMBOL_WARNING" For simplier initialization, use BSP "LV_SYMBOL_WARNING);
+#endif
     lv_obj_align(label, LV_ALIGN_CENTER, 0, -30);
 
     /* Button */
