+++ components/esp_lvgl_port/idf_component.yml
@@ -1,8 +1,8 @@
-version: "1.4.0"
+version: "1.5.0"
 description: ESP LVGL port
 url: https://github.com/espressif/esp-bsp/tree/master/components/esp_lvgl_port
 dependencies:
   idf: ">=4.4"
   lvgl/lvgl:
-    version: "^8"
+    version: ">=8,<10"
     public: true
