+++ components/esp_lvgl_port/images/lvgl8/img_cursor.c
